requirements.txt
flask
flask-cors
pdfplumber
sentence-transformers
faiss-cpu
requests
python-dotenv
numpy
